This adds persistent tracking for messages and evaluation of these ^^
Interaction with the database is wrapped in the db.js file!

The delete endpoint was removed cause it was not longer needed.
Messages have a lifetime which is also saved in the db. A "normal"
message has 24H life time a message from gpp have 72H of lifetime.

This also cleans some things up ... a more thorough patch will follow